Cleaning up plotting, fixing more of force
