+++ src/quadcoil/io/__init__.py
@@ -1,6 +1,5 @@
 from .jax import *
 from .file import *
-from .plotting import *
 from .desc import *
 from .simsopt import *
 from .coil_cutting import *
